Task Engine Added
Note: The commited code was written weeks prior to this commit.  And due to the fact that there is no testing system in place, only if or no the code compiles, and considering that this thought of commitance being considered as the next step for some time, I do hereby assume that the code in it's presetn state is ready for commitance.